backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
channels==4.1.0
uvicorn==0.30.1
djangorestframework-simplejwt==5.3.1
